Create TerrainDescriptor.cs
Place holder for editing terrain features in Editor. Not needed if terrain features are defined in a data container or file (JSON), but definitely useful for prototyping.
The game objects (or prefabs) containing this component can be placed as children of Grid GameObject (containing Grid manager or TerrainBuilder) for easy finding during Grid Creation.